ULTIMATE FIX: Resolve mixed precision data type issues in FlashInfer
This commit addresses the root cause of "apply_rope_with_cos_sin_cache_inplace
failed to dispatch data type Float" by fixing mixed precision inconsistencies
across all model architectures.

🔧 Core Issues Fixed:
- Linear layer outputs defaulting to float32 despite fp16 model dtype
- Inconsistent data types between query/key tensors and cos_sin_cache
- Missing dtype conversion in projection layers

📋 Changes by Component:

1. DeepSeek Model (model_hub/deepseek.py):
   - Force linear layer outputs to model dtype: .to(self.parent_model.dtype)
   - Fixed both prefill and decode paths

2. MLA Attention (attn_hub/mla_attn.py):
   - Ensure k_b_proj/v_b_proj outputs match query_states dtype
   - Fixed both mla_prefill_attn and mla_decode_attn functions

3. LLaMA Model (model_hub/llama.py):
   - Add dtype conversion before FlashInfer RoPE: .to(self.dtype)

4. Qwen Model (model_hub/qwen.py):
   - Add dtype conversion before FlashInfer RoPE: .to(self.dtype)

🎯 Result:
All FlashInfer function calls now receive consistent fp16/bf16 tensors:
✅ position_ids: 1D format (previous fix)
✅ query_states: fp16/bf16
✅ key_states: fp16/bf16
✅ cos_sin_cache: fp16/bf16
✅ All tensors: dtype consistent

This should completely eliminate FlashInfer data type dispatch errors.

+++ model_hub/deepseek.py
@@ -185,14 +185,14 @@ class MLAAttention(nn.Module):
         try:
             # Query低秩投影
             q_compressed = self.q_a_proj(hidden_states)  # [bs, seq_len, q_lora_rank]
-            query_states = self.q_b_proj(q_compressed)  # [bs, seq_len, num_heads * head_dim]
+            query_states = self.q_b_proj(q_compressed).to(self.parent_model.dtype)  # 强制转换为模型数据类型
             
             # KV压缩投影（这是MLA的核心优化）
             compressed_kv = self.kv_a_proj(hidden_states)  # [bs, seq_len, kv_lora_rank]
             
-            # 解压缩得到K和V
-            key_states = self.k_b_proj(compressed_kv)  # [bs, seq_len, num_heads * head_dim]
-            value_states = self.v_b_proj(compressed_kv)  # [bs, seq_len, num_heads * v_head_dim]
+            # 解压缩得到K和V，并强制转换数据类型
+            key_states = self.k_b_proj(compressed_kv).to(self.parent_model.dtype)  # 强制转换为模型数据类型
+            value_states = self.v_b_proj(compressed_kv).to(self.parent_model.dtype)  # 强制转换为模型数据类型
             
             # 重塑为多头格式
             query_states = query_states.view(batch_size, seq_len, self.num_heads, self.head_dim)
@@ -284,7 +284,7 @@ class MLAAttention(nn.Module):
         
         # Query低秩投影
         q_compressed = self.q_a_proj(hidden_states)
-        query_states = self.q_b_proj(q_compressed)
+        query_states = self.q_b_proj(q_compressed).to(self.parent_model.dtype)  # 强制转换为模型数据类型
         
         # 重塑为多头格式
         query_states = query_states.view(batch_size, 1, self.num_heads, self.head_dim)
